requirements/web_requirements.txt
# Web Backend Requirements for RAG System

# FastAPI Backend Requirements (recommended)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.0

# Required for RAG Core functionality
openai>=1.0.0
supabase>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
langchain>=0.0.300
langchain-community>=0.0.20
tiktoken>=0.5.0
tqdm>=4.65.0
aiohttp>=3.8.0

# For production deployment
#gunicorn==21.2.0  # For Flask
#prometheus-client==0.19.0  # For monitoring
#redis==5.0.1  # For session management/caching

# Security (recommended for production)
#authlib==1.2.1
#python-jose[cryptography]==3.3.0

# Database (optional - for user management)
#sqlalchemy==2.0.23
#alembic==1.13.1
